Webhook/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.0
blinker==1.6.2
boto3==1.28.50
botocore==1.31.50
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
decorator==5.1.1
Flask==2.3.3
Flask-Cors==4.0.0
flask-ngrok==0.0.25
frozenlist==1.4.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.40.1
MarkupSafe==2.1.3
msgpack==1.0.5
multidict==6.0.4
numba==0.57.1
numpy==1.24.4
openai==0.28.0
packaging==23.1
platformdirs==3.10.0
pooch==1.7.0
pycparser==2.21
pydantic==2.4.0
pydantic_core==2.10.0
pydub==0.25.1
python-dateutil==2.8.2
python-dotenv==1.0.0
replicate==0.13.0
requests==2.31.0
s3transfer==0.6.2
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
soundfile==0.12.1
soxr==0.3.6
threadpoolctl==3.2.0
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==1.26.16
Werkzeug==2.3.7
yarl==1.9.2
gunicorn
azure-cognitiveservices-speech